Merge main

+++ Cloud/src/main/java/com/example/petfeedercloud/controllers/PetFeederController.java
@@ -65,10 +65,13 @@ public class PetFeederController {
     @Operation(summary = "Delete pet feeder", description = "Delete pet Feeder by its id")
     public ResponseEntity<?> deletePet(@PathVariable Long petfeederId) {
         try {
-            petFeederService.deletePetFeeder(petfeederId);
-            return ResponseEntity.ok("Petfeeder deleted successfully");
-        } catch (NotFoundException ex) {
-            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(ex.getMessage());
+            // Check if the pet feeder exists before attempting deletion
+            if (petFeederService.getPetFeederById(petfeederId)!=null) {
+                petFeederService.deletePetFeeder(petfeederId);
+                return ResponseEntity.ok("Petfeeder deleted successfully");
+            } else {
+                return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Petfeeder not found");
+            }
         } catch (Exception ex) {
             return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("An error occurred while deleting the petfeeder");
         }
@@ -83,14 +86,13 @@ public class PetFeederController {
     @Operation(summary = "Activate pet feeder", description = "Activate a pet feeder for a specific user and pet")
     public ResponseEntity<String> activatePetFeeder(
             @RequestParam(required = true) Long userId,
-            @RequestParam(required = true) Long petId,
             @RequestParam Long petFeederId) {
         try {
-            if (userId == null || userId <= 0 || petId == null || petId <= 0) {
+            if (userId == null || userId <= 0) {
                 throw new IllegalArgumentException("userId and petId are required and must be greater than 0");
             }
 
-            petFeederService.setActivePetFeeder(userId, petId, petFeederId);
+            petFeederService.setActivePetFeeder(userId, petFeederId);
             return ResponseEntity.ok("Pet feeder activated successfully");
         } catch (IllegalArgumentException ex) {
             return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(ex.getMessage());
